--- app/src/main/java/dk/sdu/weshare/API/ErrorResponse.kt
@@ -1,7 +1,0 @@
-package dk.sdu.weshare.API
-
-import com.fasterxml.jackson.annotation.JsonProperty
-
-data class ErrorResponse(
-	@JsonProperty("message") val message: String
-)
