--- app/src/main/java/dk/sdu/weshare/API/RetrofitClient.kt
@@ -1,45 +1,0 @@
-package dk.sdu.weshare.API
-
-import okhttp3.Interceptor
-import okhttp3.OkHttpClient
-import okhttp3.Response
-import retrofit2.Retrofit
-import retrofit2.converter.jackson.JacksonConverterFactory
-import java.util.*
-
-object RetrofitClient {
-	// Base URL must end in /
-	private const val BASE_URL = "http://localhost:8090/v1/"
-
-	private val okHttpClient = OkHttpClient()
-		.newBuilder()
-		.addInterceptor(AuthorizationInterceptor)
-		.addInterceptor(RequestInterceptor)
-		.build()
-
-	fun getClient(): Retrofit =
-		Retrofit.Builder()
-			.client(okHttpClient)
-			.baseUrl(BASE_URL)
-			.addConverterFactory(JacksonConverterFactory.create())
-			.build()
-}
-
-object RequestInterceptor : Interceptor {
-	override fun intercept(chain: Interceptor.Chain): Response {
-		val request = chain.request()
-		println("Outgoing request to ${request.url()}")
-		return chain.proceed(request)
-	}
-}
-
-object AuthorizationInterceptor : Interceptor {
-	override fun intercept(chain: Interceptor.Chain): Response {
-		val requestWithHeader = chain.request()
-			.newBuilder()
-			.header(
-				"Authorization", UUID.randomUUID().toString()
-			).build()
-		return chain.proceed(requestWithHeader)
-	}
-}
