--- app/src/main/java/dk/sdu/weshare/API/UserService.kt
@@ -1,39 +1,0 @@
-package dk.sdu.weshare.API
-import com.fasterxml.jackson.databind.ObjectMapper
-import dk.sdu.weshare.models.User
-import okhttp3.ResponseBody
-class UserService {
-	private val retrofit = RetrofitClient.getClient()
-	private val userApi = retrofit.create(UserApi::class.java)
-
-	fun successfulUsersResponse() {
-		val usersResponse = userApi.getUsers()
-			.execute()
-
-		val successful = usersResponse.isSuccessful
-		val httpStatusCode = usersResponse.code()
-		val httpStatusMessage = usersResponse.message()
-
-		val body: List<User>? = usersResponse.body()
-	}
-
-	fun errorUsersResponse() {
-		val usersResponse = userApi.getUsers()
-			.execute()
-
-		val errorBody: ResponseBody? = usersResponse.errorBody()
-
-		val mapper = ObjectMapper()
-		val mappedBody: ErrorResponse? = errorBody?.let { notNullErrorBody ->
-			mapper.readValue(notNullErrorBody.toString(), ErrorResponse::class.java)
-		}
-	}
-
-	fun headersUsersResponse() {
-		val usersResponse = userApi.getUsers()
-			.execute()
-
-		val headers = usersResponse.headers()
-		val customHeaderValue = headers["custom-header"]
-	}
-}
