--- app/src/main/java/dk/sdu/weshare/fakeValues/ExpenseDetailsPageProps.kt
@@ -1,7 +1,0 @@
-package dk.sdu.weshare.fakeValues
-
-data class ExpenseDetailsPageProps(
-	val groupId: Int,
-	val expenseId: String?,
-	val onSave: () -> Unit,
-)
