--- app/src/main/java/dk/sdu/weshare/fakeValues/ProfilePageProps.kt
@@ -1,6 +1,0 @@
-package dk.sdu.weshare.fakeValues
-
-data class ProfilePageProps(
-	val userId: Int,
-	val onBack: () -> Unit,
-)
