--- app/src/main/java/dk/sdu/weshare/fakeValues/UserPageProps.kt
@@ -1,14 +1,0 @@
-package dk.sdu.weshare.fakeValues
-
-import java.time.LocalDateTime
-
-data class UserPageProps (
-	val id: Int,
-	val name: String,
-	val email: String,
-	val emailVerifiedAt: LocalDateTime,
-	val createdAt: LocalDateTime,
-	val updatedAt: LocalDateTime,
-	val groupIds: List<Int>,
-	val onBack: () -> Unit,
-)
